Hyperion-Kontrolle in eigene Datei und Klasse ausgelagert, ebenso Ansteuerung über MCP Portexpander und Volumenkontrolle
